+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 All notable changes to this package will be documented in this file. The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
+## [0.3.3] - 2022-06-13
+
+- [feat] ObjectPool に ReturnAll を追加
+- [feat] SingletonMonoBehaviour に dontDestory オプションを追加
+
 ## [0.3.2] - 2022-05-19
 
 - [feat] LocalizationData に ko を追加